+++ classes/Decorator/EventDispatcher.php
@@ -3,6 +3,7 @@
 namespace Neat\Object\Decorator;
 
 use Neat\Object\Event;
+use Neat\Object\Exception\EventNotDefinedException;
 use Neat\Object\RepositoryDecorator;
 use Neat\Object\RepositoryInterface;
 use Psr\EventDispatcher\EventDispatcherInterface;
@@ -48,14 +49,38 @@ class EventDispatcher implements RepositoryInterface
      * @param class-string $event
      * @param object       $entity
      * @return void
+     * @throws EventNotDefinedException In case the event is not defined on the entity.
      */
     public function trigger(string $event, object $entity): void
     {
-        if ($class = $this->events[$event] ?? null) {
-            $instance = new $class($entity);
+        if (!isset($this->events[$event])) {
+            $entityClass = get_class($entity);
 
-            $this->dispatcher->dispatch($instance);
+            throw new EventNotDefinedException("No event defined for '{$entityClass}' named '{$event}'");
         }
+        $this->dispatch($this->events[$event], $entity);
+    }
+
+    /**
+     * Trigger event for an entity if the entity defines the event.
+     *
+     * @param string $event
+     * @param object $entity
+     * @return void
+     */
+    public function triggerIfExists(string $event, object $entity): void
+    {
+        if (!isset($this->events[$event])) {
+            return;
+        }
+        $this->dispatch($this->events[$event], $entity);
+    }
+
+    private function dispatch(string $class, object $entity): void
+    {
+        $instance = new $class($entity);
+
+        $this->dispatcher->dispatch($instance);
     }
 
     /**
@@ -63,11 +88,11 @@ class EventDispatcher implements RepositoryInterface
      */
     public function load($entity)
     {
-        $this->trigger(Event::LOADING, $entity);
+        $this->triggerIfExists(Event::LOADING, $entity);
 
         $result = $this->repository()->load($entity);
 
-        $this->trigger(Event::LOADED, $result);
+        $this->triggerIfExists(Event::LOADED, $result);
 
         return $result;
     }
@@ -81,17 +106,17 @@ class EventDispatcher implements RepositoryInterface
         $exists     = $identifier && array_filter($identifier) && $this->has($identifier);
 
         if ($exists) {
-            $this->trigger(Event::UPDATING, $entity);
+            $this->triggerIfExists(Event::UPDATING, $entity);
         } else {
-            $this->trigger(Event::CREATING, $entity);
+            $this->triggerIfExists(Event::CREATING, $entity);
         }
 
         $this->repository()->store($entity);
 
         if ($exists) {
-            $this->trigger(Event::UPDATED, $entity);
+            $this->triggerIfExists(Event::UPDATED, $entity);
         } else {
-            $this->trigger(Event::CREATED, $entity);
+            $this->triggerIfExists(Event::CREATED, $entity);
         }
     }
 
@@ -100,11 +125,11 @@ class EventDispatcher implements RepositoryInterface
      */
     public function delete($entity)
     {
-        $this->trigger(Event::DELETING, $entity);
+        $this->triggerIfExists(Event::DELETING, $entity);
 
         $deleted = $this->repository()->delete($entity);
 
-        $this->trigger(Event::DELETED, $entity);
+        $this->triggerIfExists(Event::DELETED, $entity);
 
         return $deleted;
     }