+++ classes/Exception/EventNotDefinedException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Neat\Object\Exception;
+
+use LogicException;
+
+class EventNotDefinedException extends LogicException
+{
+
+}