+++ tests/Helper/Event/Custom.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Neat\Object\Test\Helper\Event;
+
+use Neat\Object\Event;
+
+class Custom extends Event
+{
+
+}